feat: combine route handlers and route modules
(cherry picked from commit d18539eb0130b8f1aad78acb93df54abf9d06a4c)
