--- packages/next/src/server/future/route-handlers/app-page-route-handler.ts
@@ -1,7 +1,0 @@
-import type { RouteHandler } from './route-handler'
-
-export class AppPageRouteHandler implements RouteHandler {
-  public async handle(): Promise<Response> {
-    throw new Error('Method not implemented.')
-  }
-}
