--- packages/next/src/server/future/route-handlers/pages-route-handler.ts
@@ -1,7 +1,0 @@
-import { RouteHandler } from './route-handler'
-
-export class PagesRouteHandler implements RouteHandler {
-  public async handle(): Promise<Response> {
-    throw new Error('Method not implemented.')
-  }
-}
