+++ packages/next/src/server/web/adapter.ts
@@ -51,11 +51,15 @@ const FLIGHT_PARAMETERS = [
   [FETCH_CACHE_HEADER],
 ] as const
 
-export async function adapter(params: {
+export type AdapterOptions = {
   handler: NextMiddleware
   page: string
   request: RequestData
-}): Promise<FetchEventResult> {
+}
+
+export async function adapter(
+  params: AdapterOptions
+): Promise<FetchEventResult> {
   // TODO-APP: use explicit marker for this
   const isEdgeRendering = typeof self.__BUILD_MANIFEST !== 'undefined'
